Release 1.0.0.32 QCACLD WLAN Driver

* origin/caf/caf-wlan/master:
  Cafstaging Release 1.0.0.32
  qcacld: Fixing crashes caused by dangling vdev pointer
  wlan: MAC Address Logging Fix.
  qca-cld: translation of CoEx frequency range reportings
  Kbuild: add LTE CoEx feature dependency to platform driver
  qcacld: Fix of kernel panic in PCIe interrupt handler
  qcacld: ipa: Modify IPA header file path
  wlan: make cfg.dat file read 64 bit compatible
  wlan: Multiple TX PAUSE type event handle
  wlan: TL: removing newlines from the debug logs
  wlan: WDA: removing newlines from the debug logs
  wlan: VOSS: removing newlines from the debug logs
  wlan: MAC Address Logging Fix.

+++ CORE/SYS/common/src/wlan_qct_sys.c
@@ -199,8 +199,8 @@ typedef struct sPolFileHeader
 {
   tPolFileVersion FileVersion;
   tPolFileVersion HWCapabilities;
-  unsigned long   FileLength;
-  unsigned long   NumDirectoryEntries;
+  unsigned int   FileLength;
+  unsigned int   NumDirectoryEntries;
 
 } tPolFileHeader;
 
@@ -219,9 +219,9 @@ typedef enum ePolFileDirTypes
 
 typedef struct sPolFileDirEntry
 {
-  unsigned long DirEntryType;
-  unsigned long DirEntryFileOffset;
-  unsigned long DirEntryLength;
+  unsigned int DirEntryType;
+  unsigned int DirEntryFileOffset;
+  unsigned int DirEntryLength;
 
 } tPolFileDirEntry;
